soc: qcom: glink_pkt: Increase default open timeout to 120s
Clients using the default timeout of 1 second may timeout and retry an
open operation during a critical segment of a remote subsystem boot
which can trigger a race condition and result in a system crash.

As a workaround, increase the default timeout to 120 seconds which
ensures timeouts miss the race window.

Change-Id: I7983218e3ba1bdb8c50b685e7b215b05f1142331
Signed-off-by: Steven Cahail <scahail@codeaurora.org>